@@ -33,7 +33,7 @@ include = [
 vpype = "vpype_cli.cli:cli"
 
 [tool.poetry.dependencies]
-python = ">=3.8, <3.11"
+python = ">=3.9, <3.12"
 cachetools = ">=4.2.2"
 click = ">=8.0.1,<8.2.0"
 multiprocess = ">=0.70.11"
@@ -51,7 +51,7 @@ matplotlib = { version = ">=3.3.2", optional = true }
 glcontext = { version = ">=2.3.2", optional = true }  # 2.3.2 needed to fix #200
 moderngl = { version = ">=5.6.2,!=5.7.1,!=5.7.2", optional = true }  # see moderngl/moderngl#525
 Pillow = { version = ">=9.0.1", optional = true }
-PySide6 = { version = ">=6.3.2,<6.4.0", optional = true }  # 6.4.0 incompatible with matplotlib
+PySide6 = { version = ">=6.4.0.1", optional = true }
 
 
 [tool.poetry.group.dev.dependencies]
@@ -68,6 +68,7 @@ pytest-mpl = ">=0.12"
 mypy = ">=0.901"
 types-cachetools = ">=4.2.4"
 types-setuptools = ">=57.4.17"
+types-pillow = "^9.4.0.17"
 
 [tool.poetry.group.docs]
 optional = true
