@@ -21,7 +21,7 @@ def load_icon(path: str) -> QIcon:
     if app is None or not isinstance(app, QGuiApplication):
         raise EnvironmentError("no Qt application available")
 
-    base_color = app.palette().color(QPalette.Base)
+    base_color = app.palette().color(QPalette.ColorRole.Base)
     if base_color.lightnessF() < 0.5:
         file, ext = os.path.splitext(path)
         path = file + "-dark" + ext
